+++ lib/features/digital_guide/tabs/transportation/presentation/transportation_expansion_tile_content.dart
@@ -4,9 +4,9 @@ import "package:flutter_riverpod/flutter_riverpod.dart";
 import "../../../../../config/ui_config.dart";
 import "../../../../../utils/context_extensions.dart";
 import "../../../../../widgets/my_error_widget.dart";
-import "../../../../digital_guide/data/models/digital_guide_response.dart";
-import "../../../../digital_guide/presentation/widgets/digital_guide_nav_link.dart";
 import "../../../../navigator/utils/navigation_commands.dart";
+import "../../../data/models/digital_guide_response.dart";
+import "../../../presentation/widgets/digital_guide_nav_link.dart";
 import "../data/models/digital_guide_transportation.dart";
 import "../data/repository/transportation_repository.dart";
